Fixed: Structure of the project is now understandable - Deleted unused files. - Deleted unused code - Ordered the files
