--- src/tt_api/src/components/Slider.js
@@ -1,35 +1,0 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Typography from '@material-ui/core/Typography';
-import Slider from '@material-ui/core/Slider';
-
-const useStyles = makeStyles({
-  root: {
-    width: 300,
-  },
-});
-
-function valuetext(value) {
-  return `${value}`;
-}
-
-function DiscreteSlider() {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.root}>
-      <Slider
-        defaultValue={50}
-        getAriaValueText={valuetext}
-        aria-labelledby="discrete-slider"
-        valueLabelDisplay="auto"
-        step={25}
-        marks
-        min={0}
-        max={100}
-      />
-    </div>
-  );
-}
-
-export default DiscreteSlider;
