--- src/tt_api/src/components/pages/About.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import '../../App.css';
-import Footer from '../Footer';
-import './About.css'
-
-function About () {
-    return (
-        <>
-            <div className='about-container'>
-                <h1>¿Qué es Lyrics.ia?</h1>
-                <p>Es un servicio gratuito el cual ofrece a los usuarios la posibilidad de generar nuevas letras musicales,</p>
-                <p> haciendo uso de la inteligencia artificial y una base de datos de cerca de 23,000 letras de canciones de distintos artistas.</p>
-                <p>Este servicio fue desarrollado por alumnos de la Escuela Superior de Cómputo del Instituto Politécnico Nacional.</p>
-            </div>
-            <Footer />
-        </>
-    );
-}
-
-export default About;
