--- src/tt_api/src/components/pages/Dev.css
@@ -1,53 +1,0 @@
-.dev-container {
-    height: 50%;
-    width: 100%;
-    display: block;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    box-shadow: inset 0 0 0 1000px rgba(0, 0, 0, 0);
-    object-fit: contain;    
-  }
-  
-  .dev-container > h1 {
-    color: #000;
-    font-size: 100px;
-    margin-top: -250px;
-  }
-  
-  .dev-container > p {
-    margin-top: 8px;
-    color: #000;
-    font-size: 32px;
-    font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande',
-      'Lucida Sans', Arial, sans-serif;
-  }
-
-  #iframe {
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    height: 100%;
-    width: 100%;
-  }
-
-  @media screen and (max-width: 960px) {
-    .dev-container > h1 {
-      font-size: 70px;
-      margin-top: -200px;
-    }
-  }
-  
-  @media screen and (max-width: 768px) {
-    .dev-container > h1 {
-      font-size: 40px;
-      margin-top: -200px;
-    }
-  
-    .dev-container > p {
-      font-size: 20px;      
-      text-align: center;
-    }
-  }
