--- src/tt_api/src/components/pages/Faq.css
@@ -1,55 +1,0 @@
-.faq-container {  
-    height: 90vh;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    box-shadow: inset 0 0 0 1000px rgba(0, 0, 0, 0);
-    object-fit: contain;
-  }
-  
-  .faq-container > h1 {
-    color: #000;
-    font-size: 70px;
-    margin-top: -200px;
-  }
-
-  .faq-container > h2 {
-      color: #000;
-      font-size: 30px;
-  }
-  
-  .faq-container > p {
-    margin-top: 8px;
-    color: #000;
-    font-size: 25px;
-    font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande',
-      'Lucida Sans', Arial, sans-serif;
-  }
-
-
-  
-  @media screen and (max-width: 960px) {
-    .faq-container > h1 {
-      font-size: 50px;
-      margin-top: -200px;
-    }
-  }
-  
-  @media screen and (max-width: 768px) {
-    .faq-container > h1 {
-      font-size: 30px;
-      margin-top: -200px;
-    }
-
-    .faq-container > h2 {
-        font-size: 20px;  
-        text-align: center;
-      }
-  
-    .faq-container > p {
-      font-size: 15px;      
-      text-align: center;
-    }
-  }
