--- src/tt_api/src/components/pages/Textos.py
@@ -1,5 +1,0 @@
-def addtext():
-    print("Goodbye World")
-
-print("Hello world")
-addtext()
