+++ src/tt_api/src/pages/css/Samples.css
@@ -1,0 +1,74 @@
+.samples-container {  
+    height: 90vh;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    box-shadow: inset 0 0 0 1000px rgba(0, 0, 0, 0);
+    object-fit: contain;
+  }
+  
+  .samples-container > h1 {
+    color: #000;
+    font-size: 100px;
+    margin-top: -300px;
+  }
+  
+  .samples-container > p {
+    margin-top: 8px;
+    color: #000;
+    font-size: 32px;
+    font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande',
+      'Lucida Sans', Arial, sans-serif;
+  }
+      
+  .table {      
+      margin-top: 50px;
+      width:100%;
+      border:1px;
+  }
+  
+  .table-header {
+      display:flex;
+      width:100%;
+      background:#000;
+      padding: 0;
+  }
+  
+  .table-row {
+      display:flex;
+      width:100%;
+      padding: 0;
+  }
+  
+  .table-data, .header__item {
+      flex: 1 1 20%;
+      text-align:center;
+      font-size: 20px;
+  }
+  
+  .header__item {
+      text-transform:uppercase;
+      color: #fff;
+  }
+  
+  @media screen and (max-width: 960px) {
+    .samples-container > h1 {
+      font-size: 70px;
+      margin-top: -200px;
+    }
+  }
+  
+  @media screen and (max-width: 768px) {
+    .samples-container > h1 {
+      font-size: 40px;
+      margin-top: -200px;
+    }
+  
+    .samples-container > p {
+      font-size: 20px;
+      text-align: center;
+    }
+  }
+
