+++ src/tt_api/src/pages/js/Home.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import '../../App.css';
+import Footer from '../Footer';
+import Lyricgenerator from '../Lyricgenerator';
+
+function Home () {
+    return (
+        <>
+            <Lyricgenerator/>
+            <Footer />
+        </>
+    );
+}
+
+export default Home;
